add just swap select
