+++ webui/src/mystore/index.ts
@@ -3,3 +3,4 @@ export * as swap from './swap'
 export * as wallet from './wallet'
 export * as user from './user'
 export * as bulletin from './bulletin'
+export * as trades from './trades'
