+++ webui/src/mystore/kline/types.ts
@@ -2,7 +2,6 @@ import { constants } from 'src/const'
 
 export enum API {
   GetKPointsForLine = `${constants.klineEndpoint}/v1/get/kpoints/for/line`,
-  GetTransactionsForLine = `${constants.klineEndpoint}/v1/get/transactions/for/line`,
 }
 
 export interface KPoint {
@@ -23,27 +22,5 @@ export interface GetKPointsForLineResponse {
   Offset: number;
   OriginalTime: number;
   TokenPairID: number;
-  Total: number;
-}
-
-export interface Transaction {
-  ID: number;
-  PoolID: number;
-  TransactionID: number;
-  TransactionType: string;
-  Owner: string;
-  AmountZeroIn: number;
-  AmountOneIn: number;
-  AmountZeroOut: number;
-  AmountOneOut: number;
-  Timestamp: number;
-}
-
-export interface GetTransactionsForLineResponse {
-  OriginalTxID: number;
-  PoolID: number;
-  Transactions: Transaction[];
-  Limit: number;
-  Offset: number;
   Total: number;
 }
