+++ webui/src/mystore/trades/index.ts
@@ -1,0 +1,2 @@
+export * from './local'
+export * from './types'
